Refactor Seeder class out to its own file
