+++ db/seeder.rb
@@ -1,0 +1,28 @@
+class Seeder
+  def initialize
+    @counter = 0
+  end
+
+  def create_if_none(klass, count = nil)
+    @counter += 1
+    plural = klass.name.pluralize
+
+    if klass.none?
+      message = ["Creating", count, plural].compact.join(" ")
+      puts "  #{@counter}. #{message}."
+      yield
+    else
+      puts "  #{@counter}. #{plural} already exist. Skipping."
+    end
+  end
+
+  def create_if_doesnt_exist(klass, attribute_name, attribute_value)
+    record = klass.find_by("#{attribute_name}": attribute_value)
+    if record.nil?
+      puts "  #{klass} with #{attribute_name} = #{attribute_value} not found, proceeding..."
+      yield
+    else
+      puts "  #{klass} with #{attribute_name} = #{attribute_value} found, skipping."
+    end
+  end
+end
